Correção para ler o campo numero do Documento na API Extrato do BB
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@35396 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrBaaS/ACBrExtratoAPIBB.pas
@@ -85,7 +85,7 @@ type
     findicadorTipoLancamento: Integer;
     fnumeroContaContrapartida: String;
     fnumeroCpfCnpjContrapartida: Integer;
-    fnumeroDocumento: Integer;
+    fnumeroDocumento: Int64;
     fnumeroLote: Integer;
     ftextoDescricaoHistorico: String;
     ftextoDvContaContrapartida: String;
@@ -107,7 +107,7 @@ type
     property dataMovimento: TDateTime read fdataMovimento write fdataMovimento;
     property codigoAgenciaOrigem: Integer read fcodigoAgenciaOrigem write fcodigoAgenciaOrigem;
     property numeroLote: Integer read fnumeroLote write fnumeroLote;
-    property numeroDocumento: Integer read fnumeroDocumento write fnumeroDocumento;
+    property numeroDocumento: Int64 read fnumeroDocumento write fnumeroDocumento;
     property codigoHistorico: String read fcodigoHistorico write fcodigoHistorico;
     property textoDescricaoHistorico: String read ftextoDescricaoHistorico write ftextoDescricaoHistorico;
     property valorLancamento: Double read fvalorLancamento write fvalorLancamento;
